:rocket: POC mPulse via UserContext
