+++ src/app/Layouts/defaultPageWrapper.jsx
@@ -5,12 +5,14 @@ import FooterContainer from '../containers/Footer';
 import ManifestContainer from '../containers/Manifest';
 import ServiceWorkerContainer from '../containers/ServiceWorker';
 import GlobalStyle from '../lib/globalStyles';
+import MPulseBeacon from '../containers/MPulseBeacon';
 
 const PageWrapper = ({ children }) => (
   <Fragment>
     <GlobalStyle />
     <ServiceWorkerContainer />
     <ManifestContainer />
+    <MPulseBeacon />
     <HeaderContainer />
     {children}
     <FooterContainer />
