+++ src/app/containers/ConsentBanner/index.canonical.jsx
@@ -1,8 +1,11 @@
-import React, { Fragment, useState, useEffect } from 'react';
+import React, { Fragment, useState, useEffect, useContext } from 'react';
 import Banner from './Banner/index.canonical';
 import consentBannerUtilities from './CanonicalLogic';
+import { UserContext } from '../../contexts/UserContext';
 
 const Canonical = () => {
+  const { updateCookiePolicy } = useContext(UserContext);
+
   const [showPrivacy, setShowPrivacyBanner] = useState(false);
   const [showCookie, setShowCookieBanner] = useState(false);
 
@@ -16,6 +19,11 @@ const Canonical = () => {
 
   useEffect(runInitial, []);
 
+  const onCookieAccept = () => {
+    cookieOnAllow();
+    updateCookiePolicy();
+  };
+
   return (
     <Fragment>
       {showPrivacy ? (
@@ -28,7 +36,7 @@ const Canonical = () => {
       {!showPrivacy && showCookie ? (
         <Banner
           type="cookie"
-          onAccept={cookieOnAllow}
+          onAccept={onCookieAccept}
           onReject={cookieOnReject}
         />
       ) : null}
