+++ src/app/containers/MPulseBeacon/index.jsx
@@ -1,13 +1,25 @@
-import React, { useContext } from 'react';
+import { useEffect, useContext } from 'react';
 import { DialContext } from '../../contexts/DialContext';
-import MPulseBeacon from '../../components/MPulseBeacon';
+import { UserContext } from '../../contexts/UserContext';
+import boomr from './boomr';
 
 const MPulseBeaconContainer = () => {
   const { mpulse } = useContext(DialContext);
-  const API_KEY = process.env.MPULSE_API_KEY;
+  const { personalisationEnabled } = useContext(UserContext);
+  const API_KEY = process.env.SIMORGH_MPULSE_API_KEY;
   const isEnabled = mpulse && API_KEY;
 
-  return isEnabled ? <MPulseBeacon apiKey={API_KEY} /> : null;
+  useEffect(() => {
+    if (isEnabled && personalisationEnabled) {
+      try {
+        boomr(API_KEY);
+      } catch (e) {
+        console.log(e);
+      }
+    }
+  }, [isEnabled, personalisationEnabled, API_KEY]);
+
+  return null;
 };
 
 export default MPulseBeaconContainer;
